requirements.txt
Django==4.0.1
djangorestframework==3.13.1
psycopg2==2.9.3
drf-spectacular==0.22.1
Pillow==9.1.0
uwsgi==2.0.20
django-prometheus==2.3.0
prometheus_client==0.11.0

